Added checks in image service to see if a watch even exists before going on

+++ WatchPlatform/WatchCollection.Domain.Services/WatchImageService.cs
@@ -4,14 +4,19 @@ using WatchCollection.Domain.Services.Interfaces;
 using WatchCollection.Storage.Interfaces;
 using WatchCollection.Domain.Model;
 using WatchCollection.Domain.Services.Mapping;
+using WatchCollection.Storage.Exceptions;
 
 namespace WatchCollection.Domain.Services;
 
-public class WatchImageService(IWatchImageRepository _watchImageRepository, IBlobStorageService _blobStorageService) : IWatchImageService
+public class WatchImageService(IWatchImageRepository _watchImageRepository, IBlobStorageService _blobStorageService, IWatchService _watchService) : IWatchImageService
 {
 
     public async Task<WatchImageResponseContract> UploadImageAsync(Guid watchId, string fileName, string contentType, long fileSize, WatchImageRequestContract contract, Stream imageStream)
     {
+        var watch = await _watchService.GetWatchById(watchId);
+        if (watch is null)
+            throw new WatchNotFoundException();
+
         var updatedFileName = Guid.NewGuid().ToString() + "_" + fileName; // Ensures unique filenames
         var imageUrl = await _blobStorageService.UploadImageAsync(updatedFileName, imageStream);
 
@@ -33,18 +38,30 @@ public class WatchImageService(IWatchImageRepository _watchImageRepository, IBlo
 
     public async Task<List<WatchImageResponseContract>> GetAllImagesByWatchIdAsync(Guid watchId)
     {
+        var watch = await _watchService.GetWatchById(watchId);
+        if (watch is null)
+            throw new WatchNotFoundException();
+
         var entities =  await _watchImageRepository.GetAllImagesByWatchIdAsync(watchId);
         return entities.Select(e => e.AsModel().AsResponseContract()).ToList();
     }
 
     public async Task DeleteImageAsync(Guid watchId, Guid imageId)
     {
+        var watch = await _watchService.GetWatchById(watchId);
+        if (watch is null)
+            throw new WatchNotFoundException();
+
         var blobUrl = await _watchImageRepository.DeleteWatchImageDataAsync(watchId, imageId);
         await _blobStorageService.DeleteImageAsync(blobUrl);
     }
 
     public async Task DeleteImagesByWatchIdAsync(Guid watchId)
     {
+        var watch = await _watchService.GetWatchById(watchId);
+        if (watch is null)
+            throw new WatchNotFoundException();
+
         var fileNames = await _watchImageRepository.DeleteImagesByWatchIdAsync(watchId);
         foreach (var filename in fileNames)
         {
